+++ main.c
@@ -36,7 +36,7 @@
 #include "wrapper.h"
 
 const char g_programName[]     = "Enhanced DUC";
-const char g_programVersion[]  = "v2.1-dev";
+const char g_programVersion[]  = "v2.1";
 const char g_programAuthor[]   = "Markus Uhlin";
 const char g_maintainerEmail[] = "markus.uhlin@bredband.net";