+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - **Update hostname bug due to an insufficient approach of storing the
   server response**!
 - Misspelled words
+- Non ANSI function declarations
 
 ## [2.0.1] - 2018-04-14 ##
 ### Added ###